[댓글 기능] 히브리 미션 제출합니다. (#122)
* feat: 실습

* feat: 환경설정

* feat: HTML 중복제거

* refactor: thymeleaf-layout 제거

* feat: static resource path 수정

* fix: devtools 개발 시에만 사용하도록 변경

* feat: 게시물 생성

* feat: 게시글 생성 후 전체 리스트 조회

* feat: 게시글 개별 조회

* style: formatting

* feat: Article에 id 추가

* feat: 게시글 수정 페이지로 이동

* feat: 게시글 수정

* feat: 게시글 삭제

* test: 테스트간 영향이 없도록 수정

* fix: UI 버그 수정

* style: formatting

* test: 테스트 코드 중복 제거

* test: ArticleRepositoryTests에서 서버를 띄우지 않도록 수정

* refactor: ArticleRepository 클래스의 clear 메서드 제거

* refactor: Article id의 동시성 보장

* refactor: matchId 구현

* refactor: next id를 Article에서 관리하도록 수정

* test: Article test 추가

* refactor: @Autowired를 생성자 주입으로 변경

* refactor: JAP 활용 방식으로 변경

* feat: sql formatting 설정

* test: test 코드 수정

* docs: 기능목록 작성

* refactor: matchId 메서드 삭제

* feat: DB를 H2에서 MySQL로 변경

* feat: Index 페이지에 대한 처리를 IndexController에서 담당하도록 수정

* feat: Login 페이지로 이동하는 기능 구현

UserController에서 유저와 관련된 control을 담당하도록 하였습니다.

* feat: 회원가입 페이지로 이동하는 기능 구현

* feat: 회원가입 기능 구현

* feat: email 중복 검사

* feat: email 중복 시 errorMessage를 Model에 담아서 페이지에 노출

* feat: sign-up 페이지에서 좌측 상단 로고 누르면 홈으로 이동하도록 수정

* feat: 이름 규칙 적용

이름은 2~10자로 제한하며 숫자나 특수문자가 포함될 수 없다.

* feat: 비밀번호 규칙 적용

비밀번호는 8자 이상의 소문자, 대문자, 숫자, 특수문자의 조합이다.

* feat: 비밀번호 확인 기능 구현

* feat: 유저 목록 페이지 구현

* feat: FE에서 비밀번호 검증 구현

* feat: Login 기능을 User로부터 분리

* feat: http header/body logging

* feat: Login 기능 구현

* feat: 로그인 실패 시 상황에 맞는 실패 메시지를 띄운다.

- 이메일이 없는 경우
- 비밀번호가 틀린 경우

* feat: 로그아웃 기능 구현

* feat: 로그인 한 유저가 로그인/회원가입 화면에 접근할 경우 메인 화면을 띄운다.

* feat: 마이 페이지 구현

* feat: mypage-edit으로 이동하는 기능 구현

* feat: User 이름 수정 기능 구현

* feat: LogOut 상태에서 article을 생성할 수 없도록 수정

* feat: login 상태에서만 article 작성 가능하도록 수정

* feat: 자신의 Article만 수정 가능하도록 구현

* feat: 자신의 Article만 삭제 가능하도록 구현

* feat: id에 해당하는 User의 mypage를 보여주도록 수정

* feat: 자신의 회원 정보만 수정할 수 있도록 구현

* feat: 회원 탈퇴기능 구현

* refactor: Optional.get() 사용 부분 변경

* feat: User 이름 수정시 유효성 검사

* refactor: User 이름 수정시 유효성 검사 예외처리 방법 수정

UserArgumentException class 생성

* refactor: UserRepository에서 조회 결과를 Optional로 반환하도록 수정

* test: User test 추가

- edit, delete에 대한 테스트를 추가하였습니다.
- h2 db로 변경 및 초기 데이터 추가

* style: formatting

* fix: gradle.build 오타 수정

* refactor: User의 이름, 비밀번호에 대한 유효성 검사를 service에서 domain으로 변경

* feat: User 삭제시 예외처리 추가

* refactor: Pattern을 static으로 변경

* test: LogInServiceTest 수정

* refactor: default 생성자를 private로 변경

* refactor: mysql-connector-java runtimeOnly로 변경

* refactor: ArticleService 구현

* test: LogInServiceTest 수정

* test: Controller에서 사용되는 Repository를 Service로 변경

* refactor: UserArticleMissmatchException 위치 변경

* refactor: ControllerAdvice 적용

* refactor: ControllerAdvice 추가 적용

* refactor: LogInService 수정

* refactor: updateArticle 메서드 수정

* refactor: build.gradle 수정

* refactor: ControllerAdvice를 ControllerExceptionAdvice로 이름 변경

* refactor: Dto 패키지 위치를 Service 하위로 변경

* docs: 기능 목록 추가

* refactor: user, article 패키지 분리

- 패키지 안에 관련 domain, repository 포함

* refactor: 로그인 할 때 email과 password의 검증 과정 분리

* refactor: UserDeleteException.class 메시지 수정

* refactor: User.class 메소드명 수정

- setName() -> updateName()

* refactor: 유저 일치 여부에 따른 게시글 처리 방법 수정

- 수정 페이지 접근, 게시글 수정, 게시글 삭제

* refactor: User.class <-> Article.class 양방향 맵핑

- Article.class의 userId(Long) -> author(User)로 수정
- User.class에 articles(List<Article>) 추가

* refactor: ArticleService.class 의 save 메소드 수정

* feat: comment 생성 구현

- service

* feat: Article에 등록된 모든 Comment 조회 구현

* feat: comment controller 및 view 구현

- 생성, 조회

* feat: comment update 기능 구현

- service, controller

* feat: comment delete 기능 구현

- service, controller

* feat: comment delete 기능 구현

- view

* feat: comment update View 구현

* refactor: CommentService.class deleteAll 메서드 삭제

* refactor: domain에 validate annotation 추가

* test: test 추가

- Article, User 삭제시 연관된 Comment도 삭제되는지 테스트

* refactor: data.sql 제약조건에 맞게 수정

* test: 자신이 작성한 댓글인지 확인하는 테스트 추가

* refactor: article, comment update시 author 검증로직 domain으로 이동

* refactor: CommentController에서 pathVariable로 articleId를 받지 않도록 수정

* feat: Entity에 생성일, 수정일 추가

* refactor: user 이름 수정시 @Transactional 활용

* refactor: UserController의 비교 로직 domain과 service로 분리

* refactor: 유저가 일치하지 않을 경우 UserMismatchException 반환하도록 수정

* refactor: 단방향 관계매핑으로 변경

- 삭제시 @OnDelete(action = OnDeleteAction.CASCADE) 활용

* refactor: index 페이지에서 게시글의 댓글 개수가 반영되도록 수정

* refactor: CommentCountService 분리 및 테스트 추가

* refactor: 테스트 코드 중복제거

* refactor: 테스트 코드 중복제거

* fix: 테스트 코드 실패하는 부분 수정

+++ src/main/java/techcourse/myblog/service/UserService.java
@@ -1,12 +1,12 @@
 package techcourse.myblog.service;
 
 import org.springframework.stereotype.Service;
-import techcourse.myblog.domain.User;
 import techcourse.myblog.domain.exception.UserArgumentException;
+import techcourse.myblog.domain.exception.UserMismatchException;
+import techcourse.myblog.domain.user.User;
+import techcourse.myblog.domain.user.UserRepository;
 import techcourse.myblog.service.dto.UserDto;
 import techcourse.myblog.service.dto.UserPublicInfoDto;
-import techcourse.myblog.repository.ArticleRepository;
-import techcourse.myblog.repository.UserRepository;
 import techcourse.myblog.service.exception.NotFoundUserException;
 import techcourse.myblog.service.exception.SignUpException;
 import techcourse.myblog.service.exception.UserDeleteException;
@@ -21,26 +21,28 @@ import static techcourse.myblog.domain.exception.UserArgumentException.PASSWORD_
 @Service
 public class UserService {
     private UserRepository userRepository;
-    private ArticleRepository articleRepository;
 
-    public UserService(UserRepository userRepository, ArticleRepository articleRepository) {
+    public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
-        this.articleRepository = articleRepository;
     }
 
     public List<User> findAll() {
         return userRepository.findAll();
     }
 
-    public UserPublicInfoDto findById(Long id) {
-        User user = userRepository.findById(id).orElseThrow(NotFoundUserException::new);
+    public User findById(Long id) {
+        return userRepository.findById(id).orElseThrow(NotFoundUserException::new);
+    }
+
+    public UserPublicInfoDto findUserPublicInfoById(Long id) {
+        User user = findById(id);
         return new UserPublicInfoDto(user.getId(), user.getName(), user.getEmail());
     }
 
-    public void save(UserDto userDto) {
+    public User save(UserDto userDto) {
         try {
             validate(userDto);
-            userRepository.save(userDto.toEntity());
+            return userRepository.save(userDto.toEntity());
         } catch (Exception e) {
             throw new SignUpException(e.getMessage());
         }
@@ -63,21 +65,21 @@ public class UserService {
         }
     }
 
-    public void update(UserPublicInfoDto userPublicInfoDto) {
+    @Transactional
+    public void update(UserPublicInfoDto userPublicInfoDto, Long id, Long loggedInUserId) {
         try {
-            User user = userRepository.findByEmail(userPublicInfoDto.getEmail())
-                    .orElseThrow(NotFoundUserException::new);
-            user.setName(userPublicInfoDto.getName());
-            userRepository.save(user);
+            User user = findById(id);
+            user.updateName(userPublicInfoDto.getName(), loggedInUserId);
         } catch (Exception e) {
             throw new UserUpdateException(e.getMessage());
         }
     }
 
-    @Transactional
-    public void delete(Long id) {
+    public void delete(Long id, Long loggedInUserId) {
         try {
-            articleRepository.deleteByUserId(id);
+            if (!id.equals(loggedInUserId)) {
+                throw new UserMismatchException();
+            }
             userRepository.deleteById(id);
         } catch (Exception e) {
             throw new UserDeleteException(e.getMessage());
